+++ src/world_object.cpp
@@ -1,6 +1,6 @@
 #include "world_object.hpp"
 
-WorldObject::WorldObject(double x, double y, unsigned short radius, sf::Color color)
+WorldObject::WorldObject(double x, double y, unsigned int radius, sf::Color color)
   : x(x), y(y), radius(radius) {
     appearance = sf::CircleShape(radius);
     appearance.setFillColor(color);
